wip: reworking UI css
