+++ client/src/game/game.vue
@@ -1,10 +1,7 @@
 <template>
     <div id="main" @mouseleave="mouseleave" @wheel="zoom">
-        <menu-bar v-if="showUI"></menu-bar>
+        <ui ref="ui" v-if="ready.manager"></ui>
         <div id="board">
-            <template v-if="ready.manager">
-                <tool-bar ref="tools" v-show="showUI"></tool-bar>
-            </template>
             <div
                 id="layers"
                 @mousedown="mousedown"
@@ -17,32 +14,11 @@
                 @touchstart="touchstart"
                 @touchend="touchend"
             ></div>
-            <floor-select v-show="showUI"></floor-select>
         </div>
-        <selection-info ref="selectionInfo" v-show="showUI"></selection-info>
         <initiative-dialog ref="initiative" id="initiativedialog"></initiative-dialog>
         <note-dialog ref="note"></note-dialog>
         <label-dialog ref="labels"></label-dialog>
         <dm-settings ref="dmsettings" v-if="IS_DM || FAKE_PLAYER"></dm-settings>
-        <!-- When updating zoom boundaries, also update store updateZoom function;
-        should probably do this using a store variable-->
-        <zoom-slider
-            id="zoomer"
-            v-model="zoomDisplay"
-            v-show="showUI"
-            :height="6"
-            :width="200"
-            :min="0"
-            :max="1"
-            :interval="0.1"
-            :dot-size="[8, 20]"
-            :dot-options="{ style: { 'border-radius': '15%', 'z-index': 11 } }"
-            :tooltip-placement="'bottom'"
-            :tooltip="'focus'"
-            :tooltip-formatter="zoomDisplay.toFixed(1)"
-            :rail-style="{ 'background-color': '#fff', 'box-shadow': '0.5px 0.5px 3px 1px rgba(0, 0, 0, .36)' }"
-            :process-style="{ 'background-color': '#fff' }"
-        ></zoom-slider>
         <prompt-dialog ref="prompt"></prompt-dialog>
         <confirm-dialog ref="confirm"></confirm-dialog>
     </div>
@@ -52,45 +28,34 @@
 import throttle from "lodash/throttle";
 import Vue from "vue";
 import Component from "vue-class-component";
-import vueSlider from "vue-slider-component";
-import "vue-slider-component/theme/default.css";
 
 import "@/game/api/events";
 
 import ConfirmDialog from "@/core/components/modals/confirm.vue";
 import Prompt from "@/core/components/modals/prompt.vue";
 import Initiative from "@/game/ui/initiative/initiative.vue";
-import FloorSelect from "@/game/ui/floors.vue";
 import LabelManager from "@/game/ui/labels.vue";
-import MenuBar from "@/game/ui/menu/menu.vue";
 import NoteDialog from "@/game/ui/note.vue";
-import SelectionInfo from "@/game/ui/selection/selection_info.vue";
-import Tools from "@/game/ui/tools/tools.vue";
 import DmSettings from "./ui/dmsettings.vue";
+import UI from "./ui/ui.vue";
 
 import { createConnection, socket } from "@/game/api/socket";
 import { onKeyDown, onKeyUp } from "@/game/events/keyboard";
 import { scrollZoom } from "@/game/events/mouse";
 import { layerManager } from "@/game/layers/manager";
 import { gameStore } from "@/game/store";
-import { l2g } from "@/game/units";
-import { LocalPoint } from "./geom";
 import { dropAsset } from "./layers/utils";
 import { coreStore } from "@/core/store";
 
 @Component({
     components: {
-        "tool-bar": Tools,
-        "selection-info": SelectionInfo,
         "prompt-dialog": Prompt,
         "confirm-dialog": ConfirmDialog,
-        "floor-select": FloorSelect,
-        "menu-bar": MenuBar,
         "initiative-dialog": Initiative,
-        "zoom-slider": vueSlider,
         "note-dialog": NoteDialog,
         "label-dialog": LabelManager,
         "dm-settings": DmSettings,
+        ui: UI,
     },
     beforeRouteEnter(to, from, next) {
         coreStore.setLoading(true);
@@ -107,7 +72,7 @@ export default class Game extends Vue {
         confirm: InstanceType<typeof ConfirmDialog>;
         note: InstanceType<typeof NoteDialog>;
         prompt: InstanceType<typeof Prompt>;
-        tools: InstanceType<typeof Tools>;
+        ui: InstanceType<typeof UI>;
     };
 
     ready = {
@@ -121,27 +86,12 @@ export default class Game extends Vue {
     throttledtouchmoveSet = false;
     throttledtouchmove: (event: TouchEvent) => void = (_event: TouchEvent) => {};
 
-    get showUI(): boolean {
-        return gameStore.showUI;
-    }
-
-    get IS_DM(): boolean {
-        return gameStore.IS_DM;
-    }
-
     get FAKE_PLAYER(): boolean {
         return gameStore.FAKE_PLAYER;
     }
 
-    get zoomDisplay(): number {
-        return gameStore.zoomDisplay;
-    }
-
-    set zoomDisplay(value: number) {
-        gameStore.updateZoom({
-            newZoomDisplay: value,
-            zoomLocation: l2g(new LocalPoint(window.innerWidth / 2, window.innerHeight / 2)),
-        });
+    get IS_DM(): boolean {
+        return gameStore.IS_DM;
     }
 
     mounted(): void {
@@ -173,18 +123,18 @@ export default class Game extends Vue {
     // Touch events
 
     touchend(event: TouchEvent): void {
-        this.$refs.tools.touchend(event);
+        this.$refs.ui.$refs.tools.touchend(event);
     }
 
     touchstart(event: TouchEvent): void {
-        this.$refs.tools.touchstart(event);
+        this.$refs.ui.$refs.tools.touchstart(event);
     }
 
     touchmove(event: TouchEvent): void {
         // limit the number of touch moves to ease server load
         if (!this.throttledtouchmoveSet) {
             this.throttledtouchmoveSet = true;
-            this.throttledtouchmove = throttle(this.$refs.tools.touchmove, 5);
+            this.throttledtouchmove = throttle(this.$refs.ui.$refs.tools.touchmove, 5);
         }
         // after throttling pass event to object
         this.throttledtouchmove(event);
@@ -193,27 +143,27 @@ export default class Game extends Vue {
     // Mouse events
 
     mousedown(event: MouseEvent): void {
-        this.$refs.tools.mousedown(event);
+        this.$refs.ui.$refs.tools.mousedown(event);
     }
 
     mouseup(event: MouseEvent): void {
-        this.$refs.tools.mouseup(event);
+        this.$refs.ui.$refs.tools.mouseup(event);
     }
 
     mousemove(event: MouseEvent): void {
         if (!this.throttledmoveSet) {
             this.throttledmoveSet = true;
-            this.throttledmove = throttle(this.$refs.tools.mousemove, 15);
+            this.throttledmove = throttle(this.$refs.ui.$refs.tools.mousemove, 15);
         }
         this.throttledmove(event);
     }
 
     mouseleave(event: MouseEvent): void {
-        this.$refs.tools.mouseleave(event);
+        this.$refs.ui.$refs.tools.mouseleave(event);
     }
 
     contextmenu(event: MouseEvent): void {
-        this.$refs.tools.contextmenu(event);
+        this.$refs.ui.$refs.tools.contextmenu(event);
     }
 
     async drop(event: DragEvent): Promise<void> {
@@ -266,23 +216,8 @@ svg {
 }
 
 #board {
-    position: relative;
+    position: absolute;
     width: 100%;
     height: 100%;
-}
-
-#zoomer {
-    position: absolute;
-    top: 15px;
-    right: 25px;
-    z-index: 11;
-}
-
-#FPS {
-    position: absolute;
-    top: 0;
-    right: 75px;
-    z-index: 11;
-    color: white;
 }
 </style>
